More progress on improving exports
