+++ src/helpers/icareExtraction.js
@@ -5,6 +5,7 @@ const fhirpath = require('fhirpath');
 const { logger } = require('mcode-extraction-framework');
 
 function makeUUIDFullUrl(uuid) {
+  // Utility for ensuring UUID's follow a valid URL specification as described by FHIR R4“
   return `urn:uuid:${uuid}`;
 }
 
@@ -17,6 +18,8 @@ function getResourceCountInBundle(messageBundle) {
 }
 
 function generateNewMessageBundle(bundle) {
+  // Generate a fhir message bundle structured according to the ICARE-submission procedure
+  // link: https://github.com/ICAREdata/icare-data-submission-procedures
   const siteId = fhirpath.evaluate(
     bundle,
     'Bundle.entry.where(resource.resourceType=\'ResearchStudy\').resource.site',
@@ -66,7 +69,37 @@ function generateNewMessageBundle(bundle) {
   };
 }
 
+async function extractDataForPatients(patientIds, mcodeClient, fromDate, toDate) {
+  // Using an initialized mcodeClient, extract data for patient ids in the appropriate toDate-fromDate range
+  const totalExtractionErrors = {};
+  const extractedData = [];
+  // Track if these runs were successful; if not, don't log a new RunInstance
+  let successfulExtraction = true;
+  /* eslint-disable no-restricted-syntax */
+  /* eslint-disable no-await-in-loop */
+  for (const [index, mrn] of patientIds.entries()) {
+    totalExtractionErrors[index] = [];
+    try {
+      logger.info(`Extracting information for patient at row ${index + 1} in .csv file`);
+      const { bundle, extractionErrors } = await mcodeClient.get({ mrn, fromDate, toDate });
+      totalExtractionErrors[index].push(...extractionErrors);
+      const resourceCount = getResourceCountInBundle(bundle);
+      logger.info(`Resources extracted for patient ${index + 1} in .csv file`);
+      Object.keys(resourceCount).forEach((resourceType) => logger.info(`${resourceType}: ${resourceCount[resourceType]} extracted`));
+      extractedData.push(bundle);
+    } catch (fatalErr) {
+      successfulExtraction = false;
+      totalExtractionErrors[index].push(fatalErr);
+      logger.error(`Fatal error extracting data: ${fatalErr.message}`);
+      logger.debug(fatalErr.stack);
+    }
+  }
+
+  return { extractedData, successfulExtraction, totalExtractionErrors };
+}
+
 module.exports = {
   getResourceCountInBundle,
   generateNewMessageBundle,
+  extractDataForPatients,
 };
